--- templates/admin/gestion-demandes/_demande-title.html.twig
@@ -1,9 +1,0 @@
-<div class="large-card-title">
-	<div>
-		<strong>{{ demande.user.nom }}</strong> {{ demande.user.prenom }}
-	</div>
-	<em>{{ demande.user.email}}</em>
-	<div>
-		<em style="font-size: 0.8em;">créée il y a <strong>{{ demande.createdDaysAgo() }}</strong> jour(s)</em>
-	</div>
-</div>
