+++ templates/admin/gestion-dsis/index.html.twig
@@ -15,9 +15,6 @@
 		</div>
 		<div class="subnavbar-title">
 			<h1>Gestion des fonctions DSI</h1> 
-			<p>
-				<em>Si un utilisateur spécifique n'apparaît pas, c'est sûrement car vous n'avez pas encore validé son compte. Pour ce faire, se rendre <a href="{{ path('admin.gestion-nouveaux-comptes') }}">ici</a></em>
-			</p>
 		</div>
 	</nav>
 {% endblock %}
