--- templates/admin/gestion-nouveaux-comptes/index.html.twig
@@ -1,54 +1,0 @@
-{% extends "bases/crous-base.html.twig" %}
-
-{% block subnavbar %}
-	<nav class="subnavbar-crous">
-		<div class="previous-link">
-			<a href="{{ path('home') }}">Page précédente</a>
-		</div>
-		<div class="subnavbar-title">
-			<h1>Gestion des validations de compte</h1>
-			<p>
-				<em>Pour modifier les informations liées à un agent, veuillez vous rendre <a href="{{ path('admin.gestion-utilisateurs') }}">ici</a></em>
-			</p>
-		</div>
-	</nav>
-{% endblock %}
-
-{% block body %}
-
-{% include "bases/_display-flashes.html.twig" %}
-
-<table class="table">
-	<thead>
-		<th>Identité</th>
-		<th>Services demandés</th>
-		<th>Action</th>
-	</thead>
-	<tbody>
-		{% for user in users %}
-			<tr>
-				<td><strong>{{ user.prenom }} {{ user.nom }}</strong> : <em>{{ user.email }}</em></td>
-				<td>
-					<ul>
-					{% for service in user.services %}
-						<li>{{ service.libelle_court }}</li>
-					{% endfor %}
-					</ul>
-				</td>
-				<td>
-					<a href="{{ path('security.activation', {'token': user.activation_token}) }}" class="btn btn-success" onclick="return confirm('Valider ce compte ?')">Valider</a>
-					<form action="{{ path('admin.supprimer-utilisateur', {id: user.id}) }}" style="display: inline-block;" method="POST" onsubmit="return confirm('Supprimer cette creation de compte ?')">
-						<input type="hidden" name="_token" value="{{ csrf_token('delete' ~ user.id) }}">
-						<input type="hidden" name="_method" value="DELETE">
-						<button class="btn btn-danger">Supprimer</button>
-					</form>
-				</td>
-			</tr>
-		{% else %}
-			<tr>
-				<td colspan="3">Il n'y a pas d'utilisateur a valider</td>
-			</tr>
-		{% endfor %}
-	</tbody>
-</table>
-{% endblock %}
