--- templates/admin/gestion-utilisateurs/index.html.twig
@@ -1,104 +1,0 @@
-{% extends "bases/crous-base.html.twig" %}
-
-{% block subnavbar %}
-	<nav class="subnavbar-crous">
-		<div class="previous-link">
-			<a href="{{ path('home') }}">Page précédente</a>
-		</div>
-		<div class="subnavbar-title">
-			<h1>Gestion des utilisateurs</h1> 
-			<p>
-				<em>Si un utilisateur spécifique n'apparaît pas, c'est sûrement car vous n'avez pas encore validé son compte. Pour ce faire, se rendre <a href="{{ path('admin.gestion-nouveaux-comptes') }}">ici</a></em>
-			</p>
-		</div>
-	</nav>
-{% endblock %}
-
-{% block body %}
-
-<table id="table" class="table">
-	<thead class="thead-crous">
-		<th></th>
-		<th >Nom</th>
-		<th >Prénom</th>
-		<th>Email</th>
-		<th>Fin de validité de contrat</th>
-	</thead>
-	<tbody class="tbody-crous">
-		{% for user in users %}
-		<tr>
-			<td class="vert-middle">
-				<a class="flip pencil" href="#user_detail_{{ loop.index0 }}"></a>
-				{# <form action="{{ path('admin.supprimer-utilisateur', {id: user.id}) }}" style="display: inline-block;" method="POST" onsubmit="return confirm('Supprimer cette creation de compte ?')">
-					<input type="hidden" name="_token" value="{{ csrf_token('delete' ~ user.id) }}">
-					<input type="hidden" name="_method" value="DELETE">
-					<button class="btn btn-danger">Supprimer</button>
-				</form> #}
-			</td>
-			<td class="vert-middle"><strong>{{ user.nom }}</strong></td>
-			<td class="vert-middle"><strong>{{ user.prenom }}</strong></td>
-			<td class="vert-middle">{{ user.email }}</td>
-			<td class="text-center vert-middle">{{ user.formatDate(user.getDateFinValid()) }}</td>
-		</tr>
-		{% endfor %}
-	</tbody>
-</table>
-
-{% include "bases/_display-flashes.html.twig" %}
-
-{{ form_start(form) }}
-<div class="detail-box">
-	{% for userForm in form.users %}
-	<div class="panel" id="user_detail_{{ loop.index0 }}">
-
-		<p>
-			Gestion des informations concernant l'agent <strong>{{ users[loop.index0].prenom }} {{ users[loop.index0].nom }}</strong>
-		</p>
-		<div class="row form-row-center">
-			<div class="col-md-2 label-flex">
-				<label for="users_users_{{ loop.index0 }}_nom">Nom</label>
-			</div>
-			<div id="users_users_{{ loop.index0 }}_nom" class="col-md-4 row-flex">
-				{{ form_widget(userForm.nom) }}
-			</div>
-			<div class="col-md-2 label-flex">
-				<label for="users_users_{{ loop.index0 }}_prenom">Prénom</label>
-			</div>
-			<div id="users_users_{{ loop.index0 }}_prenom" class="col-md-4 row-flex">
-				
-				{{ form_widget(userForm.prenom) }}
-			</div>
-		</div>
-		<div class="row form-row-center">
-			<div class="col-md-2 label-flex">
-				<label for="users_users_{{ loop.index0 }}_date_deb_valid">Début de contrat</label>
-			</div>
-			<div class="col-md-4 row-flex">
-				{{ form_widget(userForm.date_deb_valid) }}
-			</div>
-			<div class="col-md-2 label-flex">
-				<label for="users_users_{{ loop.index0 }}_date_fin_valid" >Fin de contrat</label>
-			</div>
-			<div class="col-md-4 row-flex">
-				{{ form_widget(userForm.date_fin_valid) }}
-			</div>
-		</div>
-		<div class="row form-row-center">
-			<div class="col-md-2 label-flex">
-				<p style="margin-right: 5px;">Services</p>
-			</div>
-			<div class="col-md-10 row-flex wide-select">
-				{{ form_widget(userForm.services) }}
-			</div>
-		</div>
-		<div id="save-div">
-			{{ form_widget(userForm.enregistrer, {'attr': {'class': 'btn bouton-crous'}}) }}
-		</div>
-
-	</div>
-	{% endfor %}
-
-</div>
-{{ form_end(form) }}
-
-{% endblock %}
