--- templates/dsi/gestion-applications/_form.html.twig
@@ -1,21 +1,0 @@
-{{ form_start(form) }}
-<div class="row">
-	<div class="col-md-4">{{ form_row(form.code, {'label': 'Code', 'attr': {'placeholder': 'code unique de l\'application'}}) }}</div>
-</div>
-<div class="row">
-	<div class="col-md-5">{{ form_row(form.libelle, {'label': 'Libellé', 'attr': {'placeholder': 'Libellé de l\'application (50 caractères)'}}) }}</div>
-	<div class="col-md-7">{{ form_row(form.libelle_long, {'label': 'Libellé long', 'attr': {'placeholder': 'Libellé long d\'une application (80 caractères)'}}) }}</div>
-</div>
-<div class="row">
-	<div class="col-md-12">
-		<input type="checkbox" name="transverse_checkbox" id="transverse_checkbox" {% if application.getType() == 't' %} checked{% endif %}>
-		<label for="transverse_checkbox">Cocher la case si c'est une application TRANSVERSE</label>
-	</div>
-</div>
-
-<hr>
-
-<div class="save-div">
-	{{ form_widget(form.save, {'label': 'Enregistrer', 'attr': {'class': 'save-button'}}) }}
-</div>
-{{ form_end(form) }}
