--- templates/dsi/gestion-applications/index.html.twig
@@ -1,60 +1,0 @@
-{% extends 'bases/crous-base.html.twig' %}
-
-{% block subnavbar %}
-	<nav class="subnavbar-crous">
-		<div class="previous-link">
-			<a href="{{ path('home') }}">Page précédente</a>
-		</div>
-		<div class="subnavbar-title">
-			<h1>Gestion des applications</h1> 
-			<p>
-				<em>Modifier, créer, supprimer une application du CROUS.</em>
-			</p>
-		</div>
-	</nav>
-{% endblock %}
-
-{% block body %}
-
-{% include "bases/_display-flashes.html.twig" %}
-
-<table class="table table-applications">
-	<thead class="thead-crous">
-		<tr>
-			<th>Code</th>
-			<th>Libellé</th>
-			<th>Transverse</th>
-			<th></th>
-		</tr>
-	</thead>
-	<tbody>
-	{% for application in applications %}
-		<tr>
-			<td><strong>{{ application.code }}</strong></td>
-			<td>{{ application.libelle }}</td>
-			<td style="text-align: center;" class="vert-middle {% if application.isTransverse() %}yes-dsi"><em>oui</em>{% else %}no-dsi"><em>non</em>{% endif %}</td>
-			<td class="actions">
-				<div class="modify-links">
-					<a class="show-link" href="{{ path('admin.gestion-applications.voir-application', {'id': application.id}) }}"></a>
-					<a class="edit-link" href="{{ path('admin.gestion-applications.editer-application', {'id': application.id}) }}"></a>
-					<form method="post" action="{{ path('admin.gestion-applications.supprimer-application', {'id': application.id}) }}" onsubmit="return confirm('Confirmation : Êtes-vous sûr de supprimer cette application ?');">
-						<input type="hidden" name="_method" value="DELETE">
-						<input type="hidden" name="_token" value="{{ csrf_token('delete' ~ application.id) }}">
-						<button class="delete-link"></button>
-					</form>
-				</div>
-			</td>
-		</tr>
-	{% else %}
-		<tr>
-			<td colspan="5">Aucune application enregistrée en base de données</td>
-		</tr>
-	{% endfor %}
-	</tbody>
-</table>
-
-<div class="add-item-div" onclick="location.href='{{ path('admin.gestion-applications.nouvelle-application') }}';" style="cursor: pointer;">
-		Créer application
-</div>
-
-{% endblock %}
