--- templates/dsi/gestion-applications/new.html.twig
@@ -1,18 +1,0 @@
-{% extends 'bases/crous-base.html.twig' %}
-
-{% block subnavbar %}
-	<nav class="subnavbar-crous">
-		<div class="previous-link">
-			<a href="{{ path('admin.gestion-applications.home') }}">Page précédente</a>
-		</div>
-		<div class="subnavbar-title">
-			<h1>Créer une nouvelle application</h1>
-		</div>
-	</nav>
-{% endblock %}
-
-{% block body %}
-
-	{% include "admin/gestion-applications/_form.html.twig" %}
-	
-{% endblock %}
