--- templates/dsi/gestion-dsis/index.html.twig
@@ -1,83 +1,0 @@
-{% import _self as formMacros %}
-{% extends "bases/crous-base.html.twig" %}
-
-{% macro printDsiForm(dsiForm) %}
-<div class="row form-row-center row-flex">
-	<div class="col-md-4" style="display: inline-flex;">de&nbsp;&nbsp;&nbsp;{{ form_widget(dsiForm.date_deb) }}</div>
-	<div class="col-md-4" style="display: inline-flex;">&nbsp;&nbsp;à&nbsp;&nbsp;&nbsp;{{ form_widget(dsiForm.date_fin) }}</div>
-</div>
-{% endmacro %}
-
-{% block subnavbar %}
-	<nav class="subnavbar-crous">
-		<div class="previous-link">
-			<a href="{{ path('home') }}">Page précédente</a>
-		</div>
-		<div class="subnavbar-title">
-			<h1>Gestion des fonctions DSI</h1> 
-			<p>
-				<em>Si un utilisateur spécifique n'apparaît pas, c'est sûrement car vous n'avez pas encore validé son compte. Pour ce faire, se rendre <a href="{{ path('admin.gestion-nouveaux-comptes') }}">ici</a></em>
-			</p>
-		</div>
-	</nav>
-{% endblock %}
-
-{% block body %}
-
-<table id="table" class="table">
-	<thead class="thead-crous">
-		<th></th>
-		<th >Nom</th>
-		<th >Prénom</th>
-		<th>Email</th>
-		<th style="text-align: center;">Actuellement DSI</th>
-	</thead>
-	<tbody>
-		{% for user in users %}
-		<tr>
-			<td class="vert-middle"><a class="flip pencil" href="#user_detail_{{ loop.index0 }}"></a></td>
-			<td class="vert-middle"><strong>{{ user.nom }}</strong></td>
-			<td class="vert-middle"><strong>{{ user.prenom }}</strong></td>
-			<td class="vert-middle">{{ user.email }}</td>
-			<td style="text-align: center;" class="vert-middle {% if user.isDSI() %}yes-dsi"><em>oui</em>{% else %}no-dsi"><em>non</em>{% endif %}</td>
-		</tr>
-		{% endfor %}
-	</tbody>
-</table>
-
-{% include "bases/_display-flashes.html.twig" %}
-
-<div class="detail-box">
-
-	{% for form in forms %}
-	<div class="panel" id="user_detail_{{ loop.index0 }}">
-
-		<div class="detail-box-banner banner_{{ loop.index0 }}">
-			<div class="detail-phrase">
-				Gestion des périodes d'accès à la fonction DSI pour l'agent <strong>{{ users[loop.index0].prenom }} {{ users[loop.index0].nom }}</strong>
-			</div>
-		</div>
-		{{ form_start(form) }}
-			<div class="form-error">
-					{{ form_errors(form)}}
-			</div>
-			<div class="form_{{ loop.index0 }}" data-prototype="{{ formMacros.printDsiForm(form.dsis.vars.prototype)|e('html_attr') }}">
-				{% for dsiForm in form.dsis %}
-					{{ formMacros.printDsiForm(dsiForm) }}
-				{% endfor %}
-			</div>
-			<hr>
-			<div class="save-div" id="save-div">
-				{{ form_widget(form.save, {'label': 'Enregistrer', 'attr': {'class': 'save-button'}}) }}
-			</div>
-			<input type="hidden" id="custId" name="custId" value="{{ loop.index0 }}">
-			<input type="hidden" id="endDate" name="endDate" value="{{ users[loop.index0].fullyFormatDate(users[loop.index0].getDateFinValid()) }}">
-		{{ form_end(form) }}
-
-	</div>
-	{% endfor %}
-
-</div>
-
-{% endblock %}
-
