+++ templates/security/access_denied.html.twig
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<title>ACCESS DENIED</title>
+	<style>
+		.main {
+			display: flex;
+			font-size: 2em;
+			font-weight: 500;
+			width: 70%;
+			text-align: center;
+			position: absolute;
+			left: 50%;
+			top: 50%;
+			transform: translate(-50%, -50%);
+		}
+	</style>
+</head>
+<body>
+	<div class="main">
+		<div>
+			<img src="/ressource/logo_Crous.png" alt="logo CROUS Clermont-Auvergne">
+		</div>
+			<div>
+				Vous essayez d'accéder à une page sans avoir les droits. L'accès vous est refusé.
+				<a href="{{ path('home') }}">Page d'accueil</a>
+		</div>
+	</div>	
+</body>
+</html>
