+++ templates/valideur/revue-droits/index.html.twig
@@ -20,7 +20,7 @@
 <div class="cards-crous">
 	{% for service in services %}
 	<div class="card">
-		<span class="notification-badge"><span>{{ service.getCouplesFinishedSoon() }}</span></span>
+		<span class="notification-badge"><span>{{ service.countFinishedSoon() }}</span></span>
 		<div class="card-body">
 			<h5 class="text-center card-title"><span>{{ service.code }}</span> : {{ service.libelle_court }}</h5>
 			<p class="card-libelle-long">{{ service.libelle_long }}</p>
